*: fix unstable test caused by TestRandomPanicAggConsume
